feat: Add Filipino language support
This commit adds support for the Filipino language by introducing a new module `fil.py` under the `sentencex/languages` directory. The module defines the `Filipino` class, which extends the `Language` base class and includes abbreviations specific to the Filipino language.

This change addresses the need to support Filipino as one of the languages in the SentenceX application.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for feature additions).